+++ pycroscopy/analysis/__init__.py
@@ -1,1 +1,3 @@
+__all__ = ['be_sho_utils']
+
 import be_sho_utils
